feat: make cdn configurable
- all parameters are configurable via angualr.json now
- cdn is optional and preconfigured for production

Issue: #9
